requirements.txt
faiss-cpu 
streamlit 
sentence-transformers 
langchain 
numpy 
packaging